rename : vercel 배포 실패로 인한 pages폴더 내 폴더명 소문자로 변경
