+++ src/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import './App.css';
-import OnBoarding from './pages/Onbaording/OnBoarding';
-import Login from './pages/Login/Login';
-import Home from './pages/Home/Home';
-import FindCrew from './pages/FindCrew/FindCrew';
+import OnBoarding from './pages/onboarding/OnBoarding';
+import Login from './pages/login/Login';
+import Home from './pages/home/Home';
+import FindCrew from './pages/findcrew/FindCrew';
 
 function App(): JSX.Element {
   return (
